requirements.txt
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-Script==2.0.6
Flask-WTF==0.15.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
Pillow==8.3.2
SQLAlchemy==1.4.21
WTForms==2.3.3
Werkzeug==2.0.1
alembic==1.6.5
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
Charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
pip==21.1.3
psycopg2-binary==2.9.1
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
setuptools==57.0.0
six==1.16.0
requests==2.26.0
wheel==0.36.2
urllib3==1.26.6